Cambio de tipos de datos en dos proyectos
Ahora los enteros de 16 bits son enteros de 64 bits tanto en «condicionales» como en «subacciones».

+++ condicionales/src/1-10.rs
@@ -1,8 +1,8 @@
 use functions::read_input;
 
 fn main() {
-    let a = read_input::<i16>("Ingrese un valor para A");
-    let b = read_input::<i16>("Ingrese un valor para B");
+    let a = read_input::<i64>("Ingrese un valor para A");
+    let b = read_input::<i64>("Ingrese un valor para B");
 
     println!("");
     if a % b == 0 {
