+++ condicionales/src/1-17.rs
@@ -1,8 +1,8 @@
 use functions::read_input;
 
 fn main() {
-    let multiplicando = read_input::<i16>("Ingrese un valor para el multiplicando");
-    let multiplicador = read_input::<i16>("Ingrese un valor para el multiplicador");
+    let multiplicando = read_input::<i64>("Ingrese un valor para el multiplicando");
+    let multiplicador = read_input::<i64>("Ingrese un valor para el multiplicador");
 
     let mut producto = 0;
