+++ subacciones/src/1-37-2.rs
@@ -1,6 +1,6 @@
 use functions::read_input;
 
-fn sumar_digitos(num: i16) -> i16 {
+fn sumar_digitos(num: i64) -> i64 {
     let centenas = num / 100;
     let decenas = (num - centenas * 100) / 10;
 	let unidades = num - (centenas * 100) - (decenas * 10);
@@ -9,7 +9,7 @@ fn sumar_digitos(num: i16) -> i16 {
 }
 
 fn main() {
-    let numero = read_input::<i16>("Introduzca un número de tres dígitos");
+    let numero = read_input::<i64>("Introduzca un número de tres dígitos");
 
     let suma_digitos = sumar_digitos(numero);
