+++ subacciones/src/1-37-3.rs
@@ -1,7 +1,7 @@
 use functions::read_input;
 use functions::draw_line;
 
-fn devolver_menor(num_a: i16, num_b: i16, num_c: i16) -> i16 {
+fn devolver_menor(num_a: i64, num_b: i64, num_c: i64) -> i64 {
     if num_a < num_b && num_a < num_c {
         num_a
     } else if num_b < num_a && num_b < num_c {
@@ -15,9 +15,9 @@ fn main() {
     let mut eleccion = 'S';
 
     while eleccion == 'S' {
-        let a = read_input::<i16>("Ingrese un valor numérico para A");
-        let b = read_input::<i16>("Ingrese un valor numérico para B");
-        let c = read_input::<i16>("Ingrese un valor numérico para C");
+        let a = read_input::<i64>("Ingrese un valor numérico para A");
+        let b = read_input::<i64>("Ingrese un valor numérico para B");
+        let c = read_input::<i64>("Ingrese un valor numérico para C");
 
         let menor = devolver_menor(a, b, c);
